requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
neo4j>=5.0.0
networkx>=3.0
python-dotenv>=1.0.0
rich>=13.0.0
